fix 3-fold repetation wrt TT hits.
A bug occurred in 6r1/8/8/1R3p2/5P2/1P5R/5k1K/8 w - - 19 72, where the
top move leads into a forced 3-fold repetation in a winning position.
The TT was hitting with PVNode type, returning a winning score, thus the
draw score was never discovered. Going deeper didn't help because the TT
always had an entry with appropriate depth that was a valid winning
score because the position for the TT entry only repeated once. When we
were hitting we already repeated multiple times unbeknown to the TT
logic.
The bug is basically the different evaluation value of a position with
the same hash depending on how many times it repeated.
The fix adds the 3-fold count to TT entries, and makes a condition on TT
lookup that it cannot be repeated less then the current position.