fix race condition on corrupting the PV
The PV was saved in the PV store and the returned PV was referencing the
array through a returned slice in the first ply of the store. In case of
an abort, this could have been corrupted. Now the iterative deepening
saves and returns just the bestmove, which is a copy.