respond with readyok immediatly during a search (#114)
UCI proto mandates that engine should respond with readyok immediatly if
a search is running

bench 11856525

This branch was tested with LTC for crashes. 

https://paulsonkoly.pythonanywhere.com/test/249/

In banksia we got LTC crashes, which were thought to be caused by not responding to isready. Other UIs (cutechess) don't have this issue. This is a speculative fix.

Either way we were not obeying UCI protocol.