remove checksum verification from chunks (#159)
This was stalling the server too much. While there would be options to
make it faster, for instance have a separate thread produce serverChunks
and the scheduler read a queue not doing the check-summing itself, as
the server was limited by checksumming not actual grad calcs, the best
we could hope for is the batchtime shrinking to checksumming time.

While this verification step is nice to have it is not essential and it
is too costly.

I thought about making it optional driven by cmd line flags, but
honestly if that's a choice between run fast vs run slow then it would
be just dead code. Noone ever choses run slow.

This commit removed all associated code to these checksum calcs.

bench 10918860