fix timeout invalid move
when there is a time out before we were able to produce a single move,
we printed the null move as a reposnse. This is taken as invalid move by
the UIs and the engine is killed, also depending on flags and UI
tournament stops. We are losing on time anyway. We stop printing null
move. So this case bacomes a lost game instead of an abort of
tournament.

exempt from sprt

bench 10918860