fix timeout invalid move (#163)
fix invalid move (0000) on timeout

When there was a time out before we were able to produce a single move,
we printed the null move as a reposnse. This was taken as invalid move by
the UIs and the engine was killed, also depending on flags and UI
tournament stopped. 

This is because in the ID we were interrupted before we finished depth 0, which doesn not produce a move.  Now we ignore the interrupt on depth 0.

hyper bullet

Elo   | 11.02 +- 8.09 (95%)
SPRT  | 1.0+0.01s Threads=1 Hash=4MB
LLR   | 2.95 (-2.94, 2.94) [-5.00, 0.00]
Games | N: 4258 W: 1485 L: 1350 D: 1423
Penta | [190, 415, 820, 478, 226]
https://paulsonkoly.pythonanywhere.com/test/386/

STC

Elo   | -0.22 +- 2.71 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=4MB
LLR   | 2.95 (-2.94, 2.94) [-5.00, 0.00]
Games | N: 32526 W: 9799 L: 9820 D: 12907
Penta | [1126, 3880, 6272, 3859, 1126]
https://paulsonkoly.pythonanywhere.com/test/385/

bench 10918860