board hash refactor
This is motivated by the optimised ParseFEN vs the original FromFEN
functions deviating in behaviour. For tuner tools the optimised
allocation free ParseFEN is useful, but we need to be able to set the
Hash on the board in case we need to execute moves on the board.

exempt from SPRT

bench 10918860