timeout fix attempt no.2
The previous fix attempt (cdabb449) at null moves in hard timeout was
not adequate. We still see the sporadic null move errors in
hyper-bullet. This is a more focused fix that directly targets making
the move not null if timeout happens.

This also alters behaviour, we ditch the previous solutions attempt to
use abort search result, which was guarded by
1.) being in aspiration window
2.) scoreSample > score

This is probably an incorrect thing to do anyway, because we have no
guarantees on the depth that was reached, so the quality if the aborted
search can be anything.

bench 10918860