timeout fix attempt no.2 (#167)
The previous fix attempt (cdabb449) at null moves in hard timeout was
not adequate. We still see the sporadic null move errors in
hyper-bullet. This is a more focused fix that directly targets making
the move not null if timeout happens.

This also alters behaviour, we ditch the previous solutions attempt to
use abort search result, which was guarded by
1.) being in aspiration window
2.) scoreSample > score

This was probably an incorrect thing to do anyway, because we had no
guarantees on the depth that was reached, so the quality if the aborted
search can be anything.

Increased hard timeout factor from 3 to 4, which made eliminated the elo regression.

Elo   | 7.88 +- 6.64 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=4MB
LLR   | 2.95 (-2.94, 2.94) [-5.00, 0.00]
Games | N: 5202 W: 1583 L: 1465 D: 2154
Penta | [155, 579, 1042, 643, 182]
https://paulsonkoly.pythonanywhere.com/test/392/

bench 10918860