timeout fix attempt #3 (#170)
After 74195500 the sporadic timeouts are still happening, probably
exacerbated by 63de0d0f. The fix in the previous commit is good, just
doesn't cover all cases. The tt change can cause us to return without a
move with much higher probability if the search kept hitting the tt
during the timeout. If the tt entry does not contain a move, even on the
soft timeout path we returned with a null move.

This change removes the unnecessary check for abort on the soft timeout
path, as abort should imply a soft timeout as well. It also adds a check
for move being 0, and if so we don't return from the search.

Elo   | 0.98 +- 3.51 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=4MB
LLR   | 2.98 (-2.94, 2.94) [-5.00, 0.00]
Games | N: 18478 W: 5344 L: 5292 D: 7842
Penta | [586, 2160, 3689, 2224, 580]
https://paulsonkoly.pythonanywhere.com/test/404/

bench 10385072