remove setting pv to null in nmp
This is not needed, the comment was wrong. It should be a no-change, as
the code wasn't actually doing anything. The pv couldn't have a non-null
move, even if the nmp left a pv fragment on the pv buffer, that fragment
would have started on the next ply. This is because we pass ply+1 to the
recursive call.

bench 10903418