go nodes
Hard node limit implementation. We go to exact node counts. The last
aborted depth might not influence the search result. It should be
possible to replay a pgn in the engine based on reported node counts,
assuming that the internal states at the beginning match. This can be
forced by ucinewgame before games and before the playback. This is
useful for debugging openbench crashes.

We don't yet support soft node counts.

bench number change can be attributed to what nodes we include in node
counting. Previously for instance 3 fold reps and tt cuts weren't
counted.

bench 11992731