go nodes (#178)
Hard node limit implementation. We go to exact node counts. The last
aborted depth might not influence the search result. It should be
possible to replay a pgn in the engine based on reported node counts,
assuming that the internal states at the beginning match. This can be
forced by ucinewgame before games and before the playback. This is
useful for debugging openbench crashes.

We don't yet support soft node counts.

bench number change can be attributed to what nodes we include in node
counting. Previously for instance 3 fold reps and tt cuts weren't
counted.

* added ttcut to qsearch

The SPRT regression has to be attributed to the new implementation
omitting tt cuts at depth 0 and transitioning into qsearch. The way node
counting is done now it's hard to alter the alpha/beta logic to account
the node to A/B if it's a tt cut and qsearch if not. Instead we now tt
cut in qsearch - although we don't update the tt in qsearch (yet).

Elo   | 5.70 +- 5.76 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=1MB
LLR   | 2.95 (-2.94, 2.94) [-5.00, 0.00]
Games | N: 6588 W: 1946 L: 1838 D: 2804
Penta | [172, 784, 1300, 840, 198]
https://paulsonkoly.pythonanywhere.com/test/425/

bench 11861936